Revert "Merge branch 'tharusha' into enableJWT"
This reverts commit b65fa72225076354fcf89c0eea6cf597429cfc46, reversing
changes made to 37b2ecfd4c9c5bc12234912579059a3e131ee0e7.

+++ src/main/java/com/eQuor/backend/repositories/StudentRepository.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @RepositoryRestResource(collectionResourceRel = "student", path = "student")
 @CrossOrigin
-public interface StudentRepository extends JpaRepository<Student,Long> {
+public interface StudentRepository extends JpaRepository<Student,Integer> {
     Student findByUsername(String username);
-
 }
